requirements.txt
numpy       ~= 1.19.5
matplotlib  ~= 3.3.3
PySide2     ~= 5.15.2
pytest-qt   ~= 3.3.0